Fix research sometimes not getting researched after clicking button
Caused by some kind of a race condition.

I'm not sure why selection and performing the research were separate packets.
My best guess is that this to allow the server to do checks, but the server is clearly not giving the client any feedback, so simplifying at least for now.